feat(matrix): change dtos to get student matrix

+++ src/main/resources/db/migration/V1.20250206194658__CREATE_COMPONENTS_TABLE.sql
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS components
     description                  TEXT ,
     equivalent                TEXT ,
     component_id              INTEGER ,
-    curricular_matrix_id      INTEGER ,
+    curricular_matrix_id      BIGINT ,
     name                      VARCHAR(255) ,
     pre_requisites            TEXT ,
     created_at                TIMESTAMPTZ DEFAULT NOW(),
